changed Place class into enum
