+++ src/main/java/org/jjm/destination/Destination.java
@@ -2,6 +2,7 @@ package org.jjm.destination;
 
 import org.jjm.destination.activitiy.Activity;
 import org.jjm.enums.City;
+import org.jjm.enums.Place;
 import org.jjm.enums.TransportType;
 import org.jjm.exceptions.NoActivityException;
 import org.jjm.exceptions.NoTransportException;
